Clean up of PSD class
This is the first phase of cleaning up the PSD class.
Changed it to not compute dataY's FFT if dataX is the same as dataY.
Renamed poorly named fields
Changed the segment copying to use Arrays.copyOfRange, it shaves 1
second from the unit tests; meaning more during actual run time.