Remove hard codings from Strong Motion metric
The layout of the metric has changed a little. It now uses the existing
rotated data. This is filtered and integrated to displacement. This is
different from the previous filtering and integrating non rotated data
and then rotating the filtered data.

Also it has changed from ignoring anticorrelated data to allowing <-0.85
Pearson's correlations as well as >0.85. This changed the expected
results in the test case, as NWAO 10-LHED now is computed.