Add Vacuum Monitor Metric
Add metric
It returns the units in kPa. It also compares stage0 to verify the
gains of stage 1 and 2.
The metric is able to be customized as to which channels are used for
the metric.

It does not check for being a pole zero. Just trusts that the user has
specified an actual vacuum channel.

Add a unit test
It tests IU ADK.

Add test data for IU ADK 2015-040