Remove unneeded exceptions thrown by MemberDigest
Every instance was simply being logged and ignored.
CloneNotSupported didn't make sense in this situation and was not actually thrown anywhere.
The generic RuntimeException is a bad practice as the code is now
sprinkled with try/catches for RuntimeException that do nothing and
prevent important exceptions from propagating.

I have removed the throws method declarations to be more clear. If an
Algorithm is specified that is not available, that is on the programmer
to discover not on Seedscan to catch and ignore. It should crash
spectacularly.