Commented out some risky OO code and left a rational
Code in Sequence was assuming that "this" was a Dataset. Can't assume
that with a parent class. I commented the section out since it was just
a debug statement. Added a note explaining why next to Mike's note
saying "MTH: 2013-04-27 This seems wrong:".  Not going to delete the
dead code since I may need to debug what Mike spotted as wrong.