Refactor existing Time objects to Java 8 classes
This was initially caused by the deprecation of Calendar in the jdbc
drivers and their replacement with LocalDate. I started out thinking it
was going to be a quick fix to swap out references, got a little out of
control and replaced every Calendar or GregorianCalendar reference with
either LocalDate or LocalDateTime as needed.

All Temporals in Seedscan are UTC. Tested against known data and the
results are the same. No known regressions were introduced. I went line
by line to verify all changes prior to this commit.

I also fixed numerous logger formatting issues and any general code that
appeared to need tweaking or improvement.